refactor faustgen_factory class to separate files
